ASTの要素クラスをjig/collector/domain以下に移動
